Split the models across relevant files
