+++ backend/app/models/__init__.py
@@ -1,0 +1,14 @@
+from .auth import Token, TokenPayload
+from .item import Item, ItemCreate, ItemPublic, ItemsPublic, ItemUpdate
+from .message import Message
+from .user import (
+    User,
+    UserCreate,
+    UserPublic,
+    UserRegister,
+    UsersPublic,
+    UserUpdate,
+    UserUpdateMe,
+    NewPassword,
+    UpdatePassword,
+)
