+++ backend/app/models/auth.py
@@ -1,0 +1,12 @@
+from sqlmodel import Field, SQLModel
+
+
+# JSON payload containing access token
+class Token(SQLModel):
+    access_token: str
+    token_type: str = "bearer"
+
+
+# Contents of JWT token
+class TokenPayload(SQLModel):
+    sub: str | None = None
